+++ src/node_quic_stream.h
@@ -152,9 +152,9 @@ class QuicStream : public AsyncWrap,
 
   virtual void AckedDataOffset(uint64_t offset, size_t datalen);
 
-  virtual void Close(uint16_t app_error_code = 0);
+  virtual void Close(uint64_t app_error_code = 0);
 
-  virtual void Reset(uint64_t final_size, uint16_t app_error_code = 0);
+  virtual void Reset(uint64_t final_size, uint64_t app_error_code = 0);
 
   virtual void Destroy();
